+++ controllers/home_test.go
@@ -8,17 +8,19 @@ import (
 
 	"github.com/gutakk/go-google-scraper/db"
 	"github.com/gutakk/go-google-scraper/models"
-	"github.com/gutakk/go-google-scraper/tests"
+	testConfig "github.com/gutakk/go-google-scraper/tests/config"
+	testDB "github.com/gutakk/go-google-scraper/tests/db"
+	testHttp "github.com/gutakk/go-google-scraper/tests/http"
 	"gopkg.in/go-playground/assert.v1"
 	"gorm.io/driver/postgres"
 	"gorm.io/gorm"
 )
 
 func TestDisplayHomeWithGuestUser(t *testing.T) {
-	engine := tests.GetRouter(true)
+	engine := testConfig.GetRouter(true)
 	new(HomeController).applyRoutes(engine)
 
-	w := tests.PerformRequest(engine, "GET", "/", nil, nil)
+	w := testHttp.PerformRequest(engine, "GET", "/", nil, nil)
 	p, err := ioutil.ReadAll(w.Body)
 	isHomePage := err == nil && strings.Index(string(p), "<title>Home</title>") > 0
 
@@ -27,14 +29,14 @@ func TestDisplayHomeWithGuestUser(t *testing.T) {
 }
 
 func TestDisplayHomeWithAuthenticatedUser(t *testing.T) {
-	testDB, _ := gorm.Open(postgres.Open(tests.ConstructTestDsn()), &gorm.Config{})
+	testDB, _ := gorm.Open(postgres.Open(testDB.ConstructTestDsn()), &gorm.Config{})
 	db.GetDB = func() *gorm.DB {
 		return testDB
 	}
 
 	_ = db.GetDB().AutoMigrate(&models.User{})
 
-	engine := tests.GetRouter(true)
+	engine := testConfig.GetRouter(true)
 	new(HomeController).applyRoutes(engine)
 
 	// Cookie from login API Set-Cookie header
@@ -42,7 +44,7 @@ func TestDisplayHomeWithAuthenticatedUser(t *testing.T) {
 	cookie := "go-google-scraper=MTYwNjQ2Mjk3MXxEdi1CQkFFQ180SUFBUkFCRUFBQUlmLUNBQUVHYzNSeWFXNW5EQWtBQjNWelpYSmZhV1FFZFdsdWRBWUVBUDRFdFE9PXzl6APqAQw3gAQqlHoXMYrPpnqPFkEP8SRHJZEpl-_LDQ=="
 	headers.Set("Cookie", cookie)
 
-	response := tests.PerformRequest(engine, "GET", "/", headers, nil)
+	response := testHttp.PerformRequest(engine, "GET", "/", headers, nil)
 	p, err := ioutil.ReadAll(response.Body)
 	isHomePage := err == nil && strings.Index(string(p), "<title>Home</title>") > 0
