+++ controllers/keyword_test.go
@@ -19,7 +19,9 @@ import (
 	"github.com/gin-gonic/gin"
 	"github.com/gutakk/go-google-scraper/db"
 	"github.com/gutakk/go-google-scraper/models"
-	"github.com/gutakk/go-google-scraper/tests"
+	testConfig "github.com/gutakk/go-google-scraper/tests/config"
+	testDB "github.com/gutakk/go-google-scraper/tests/db"
+	testHttp "github.com/gutakk/go-google-scraper/tests/http"
 	"github.com/stretchr/testify/suite"
 	"golang.org/x/crypto/bcrypt"
 	"gopkg.in/go-playground/assert.v1"
@@ -69,14 +71,14 @@ type KeywordDbTestSuite struct {
 }
 
 func (s *KeywordDbTestSuite) SetupTest() {
-	testDB, _ := gorm.Open(postgres.Open(tests.ConstructTestDsn()), &gorm.Config{})
+	testDB, _ := gorm.Open(postgres.Open(testDB.ConstructTestDsn()), &gorm.Config{})
 	db.GetDB = func() *gorm.DB {
 		return testDB
 	}
 
 	_ = db.GetDB().AutoMigrate(&models.User{}, &models.Keyword{})
 
-	s.engine = tests.GetRouter(true)
+	s.engine = testConfig.GetRouter(true)
 	new(LoginController).applyRoutes(EnsureGuestUserGroup(s.engine))
 	new(KeywordController).applyRoutes(EnsureAuthenticatedUserGroup(s.engine))
 
@@ -93,7 +95,7 @@ func (s *KeywordDbTestSuite) SetupTest() {
 
 	headers := http.Header{}
 	headers.Set("Content-Type", "application/x-www-form-urlencoded")
-	response := tests.PerformRequest(s.engine, "POST", "/login", headers, formData)
+	response := testHttp.PerformRequest(s.engine, "POST", "/login", headers, formData)
 
 	s.cookie = response.Header().Get("Set-Cookie")
 }
@@ -138,27 +140,27 @@ func (s *KeywordDbTestSuite) TestUploadKeywordWithBlankPayload() {
 }
 
 func TestDisplayKeywordWithGuestUser(t *testing.T) {
-	engine := tests.GetRouter(true)
+	engine := testConfig.GetRouter(true)
 	new(KeywordController).applyRoutes(EnsureAuthenticatedUserGroup(engine))
 
-	response := tests.PerformRequest(engine, "GET", "/keyword", nil, nil)
+	response := testHttp.PerformRequest(engine, "GET", "/keyword", nil, nil)
 
 	assert.Equal(t, http.StatusFound, response.Code)
 	assert.Equal(t, "/login", response.Header().Get("Location"))
 }
 
 func TestUploadKeywordWithGuestUser(t *testing.T) {
-	engine := tests.GetRouter(true)
+	engine := testConfig.GetRouter(true)
 	new(KeywordController).applyRoutes(EnsureAuthenticatedUserGroup(engine))
 
-	response := tests.PerformRequest(engine, "POST", "/keyword", nil, nil)
+	response := testHttp.PerformRequest(engine, "POST", "/keyword", nil, nil)
 
 	assert.Equal(t, http.StatusFound, response.Code)
 	assert.Equal(t, "/login", response.Header().Get("Location"))
 }
 
 func TestDisplayKeywordWithAuthenticatedUser(t *testing.T) {
-	engine := tests.GetRouter(true)
+	engine := testConfig.GetRouter(true)
 	new(KeywordController).applyRoutes(EnsureAuthenticatedUserGroup(engine))
 
 	// Cookie from login API Set-Cookie header
@@ -166,7 +168,7 @@ func TestDisplayKeywordWithAuthenticatedUser(t *testing.T) {
 	cookie := "go-google-scraper=MTYwNjQ2Mjk3MXxEdi1CQkFFQ180SUFBUkFCRUFBQUlmLUNBQUVHYzNSeWFXNW5EQWtBQjNWelpYSmZhV1FFZFdsdWRBWUVBUDRFdFE9PXzl6APqAQw3gAQqlHoXMYrPpnqPFkEP8SRHJZEpl-_LDQ=="
 	headers.Set("Cookie", cookie)
 
-	response := tests.PerformRequest(engine, "GET", "/keyword", headers, nil)
+	response := testHttp.PerformRequest(engine, "GET", "/keyword", headers, nil)
 	p, err := ioutil.ReadAll(response.Body)
 	isKeywordPage := err == nil && strings.Index(string(p), "<title>Keyword</title>") > 0
