+++ controllers/login_test.go
@@ -11,7 +11,9 @@ import (
 	"github.com/gin-gonic/gin"
 	"github.com/gutakk/go-google-scraper/db"
 	"github.com/gutakk/go-google-scraper/models"
-	"github.com/gutakk/go-google-scraper/tests"
+	testConfig "github.com/gutakk/go-google-scraper/tests/config"
+	testDB "github.com/gutakk/go-google-scraper/tests/db"
+	testHttp "github.com/gutakk/go-google-scraper/tests/http"
 	"github.com/stretchr/testify/suite"
 	"golang.org/x/crypto/bcrypt"
 	"gopkg.in/go-playground/assert.v1"
@@ -29,14 +31,14 @@ type LoginDbTestSuite struct {
 }
 
 func (s *LoginDbTestSuite) SetupTest() {
-	testDB, _ := gorm.Open(postgres.Open(tests.ConstructTestDsn()), &gorm.Config{})
+	testDB, _ := gorm.Open(postgres.Open(testDB.ConstructTestDsn()), &gorm.Config{})
 	db.GetDB = func() *gorm.DB {
 		return testDB
 	}
 
 	_ = db.GetDB().AutoMigrate(&models.User{})
 
-	s.engine = tests.GetRouter(true)
+	s.engine = testConfig.GetRouter(true)
 	new(LoginController).applyRoutes(EnsureGuestUserGroup(s.engine))
 
 	s.headers = http.Header{}
@@ -62,7 +64,7 @@ func TestLoginDbTestSuite(t *testing.T) {
 }
 
 func (s *LoginDbTestSuite) TestLoginWithValidParameters() {
-	response := tests.PerformRequest(s.engine, "POST", "/login", s.headers, s.formData)
+	response := testHttp.PerformRequest(s.engine, "POST", "/login", s.headers, s.formData)
 
 	assert.Equal(s.T(), http.StatusFound, response.Code)
 	assert.Equal(s.T(), "/", response.Header().Get("Location"))
@@ -73,7 +75,7 @@ func (s *LoginDbTestSuite) TestDisplayLoginWithAuthenticatedUser() {
 	cookie := "go-google-scraper=MTYwNjQ2Mjk3MXxEdi1CQkFFQ180SUFBUkFCRUFBQUlmLUNBQUVHYzNSeWFXNW5EQWtBQjNWelpYSmZhV1FFZFdsdWRBWUVBUDRFdFE9PXzl6APqAQw3gAQqlHoXMYrPpnqPFkEP8SRHJZEpl-_LDQ=="
 	s.headers.Set("Cookie", cookie)
 
-	response := tests.PerformRequest(s.engine, "GET", "/login", s.headers, nil)
+	response := testHttp.PerformRequest(s.engine, "GET", "/login", s.headers, nil)
 
 	assert.Equal(s.T(), http.StatusFound, response.Code)
 	assert.Equal(s.T(), "/", response.Header().Get("Location"))
@@ -82,7 +84,7 @@ func (s *LoginDbTestSuite) TestDisplayLoginWithAuthenticatedUser() {
 func (s *LoginDbTestSuite) TestLoginWithBlankEmailValidation() {
 	s.formData.Del("email")
 
-	response := tests.PerformRequest(s.engine, "POST", "/login", s.headers, s.formData)
+	response := testHttp.PerformRequest(s.engine, "POST", "/login", s.headers, s.formData)
 	p, err := ioutil.ReadAll(response.Body)
 	pageError := err == nil && strings.Index(string(p), "Invalid email format") > 0
 
@@ -93,7 +95,7 @@ func (s *LoginDbTestSuite) TestLoginWithBlankEmailValidation() {
 func (s *LoginDbTestSuite) TestLoginWithBlankPasswordValidation() {
 	s.formData.Del("password")
 
-	response := tests.PerformRequest(s.engine, "POST", "/login", s.headers, s.formData)
+	response := testHttp.PerformRequest(s.engine, "POST", "/login", s.headers, s.formData)
 	p, err := ioutil.ReadAll(response.Body)
 	pageError := err == nil && strings.Index(string(p), "Password is required") > 0
 	isEmailFieldValueExist := err == nil && strings.Index(string(p), s.email) > 0
@@ -106,7 +108,7 @@ func (s *LoginDbTestSuite) TestLoginWithBlankPasswordValidation() {
 func (s *LoginDbTestSuite) TestLoginWithTooShortPasswordValidation() {
 	s.formData.Set("password", "12345")
 
-	response := tests.PerformRequest(s.engine, "POST", "/login", s.headers, s.formData)
+	response := testHttp.PerformRequest(s.engine, "POST", "/login", s.headers, s.formData)
 	p, err := ioutil.ReadAll(response.Body)
 	pageError := err == nil && strings.Index(string(p), "Password must be longer than 6") > 0
 	isEmailFieldValueExist := err == nil && strings.Index(string(p), s.email) > 0
@@ -119,7 +121,7 @@ func (s *LoginDbTestSuite) TestLoginWithTooShortPasswordValidation() {
 func (s *LoginDbTestSuite) TestLoginWithInvalidEmail() {
 	s.formData.Set("email", "test@email.com")
 
-	response := tests.PerformRequest(s.engine, "POST", "/login", s.headers, s.formData)
+	response := testHttp.PerformRequest(s.engine, "POST", "/login", s.headers, s.formData)
 	p, err := ioutil.ReadAll(response.Body)
 	pageError := err == nil && strings.Index(string(p), "Username or password is invalid") > 0
 	isEmailFieldValueExist := err == nil && strings.Index(string(p), "test@email.com") > 0
@@ -132,7 +134,7 @@ func (s *LoginDbTestSuite) TestLoginWithInvalidEmail() {
 func (s *LoginDbTestSuite) TestLoginWithInvalidPassword() {
 	s.formData.Set("password", "123456789")
 
-	response := tests.PerformRequest(s.engine, "POST", "/login", s.headers, s.formData)
+	response := testHttp.PerformRequest(s.engine, "POST", "/login", s.headers, s.formData)
 	p, err := ioutil.ReadAll(response.Body)
 	pageError := err == nil && strings.Index(string(p), "Username or password is invalid") > 0
 	isEmailFieldValueExist := err == nil && strings.Index(string(p), s.email) > 0
@@ -143,10 +145,10 @@ func (s *LoginDbTestSuite) TestLoginWithInvalidPassword() {
 }
 
 func TestDisplayLogin(t *testing.T) {
-	engine := tests.GetRouter(true)
+	engine := testConfig.GetRouter(true)
 	new(LoginController).applyRoutes(EnsureGuestUserGroup(engine))
 
-	response := tests.PerformRequest(engine, "GET", "/login", nil, nil)
+	response := testHttp.PerformRequest(engine, "GET", "/login", nil, nil)
 	p, err := ioutil.ReadAll(response.Body)
 	pageOK := err == nil && strings.Index(string(p), "<title>Login</title>") > 0
