+++ controllers/logout_test.go
@@ -5,7 +5,8 @@ import (
 	"testing"
 
 	"github.com/gin-gonic/gin"
-	"github.com/gutakk/go-google-scraper/tests"
+	testConfig "github.com/gutakk/go-google-scraper/tests/config"
+	testHttp "github.com/gutakk/go-google-scraper/tests/http"
 	"github.com/stretchr/testify/suite"
 	"gopkg.in/go-playground/assert.v1"
 )
@@ -16,7 +17,7 @@ type LogoutTestSuite struct {
 }
 
 func (s *LogoutTestSuite) SetupTest() {
-	s.engine = tests.GetRouter(false)
+	s.engine = testConfig.GetRouter(false)
 	new(LogoutController).applyRoutes(EnsureAuthenticatedUserGroup(s.engine))
 }
 
@@ -29,14 +30,14 @@ func (s *LogoutTestSuite) TestLogoutWithAuthenticatedUser() {
 	headers := http.Header{}
 	headers.Set("Cookie", cookie)
 
-	response := tests.PerformRequest(s.engine, "POST", "/logout", headers, nil)
+	response := testHttp.PerformRequest(s.engine, "POST", "/logout", headers, nil)
 
 	assert.Equal(s.T(), http.StatusFound, response.Code)
 	assert.Equal(s.T(), "/", response.Header().Get("Location"))
 }
 
 func (s *LogoutTestSuite) TestLogoutWithGuestUser() {
-	response := tests.PerformRequest(s.engine, "POST", "/logout", nil, nil)
+	response := testHttp.PerformRequest(s.engine, "POST", "/logout", nil, nil)
 
 	// TODO: Research the flash messge assertion solution
 	assert.Equal(s.T(), http.StatusFound, response.Code)
