+++ controllers/register_test.go
@@ -11,7 +11,9 @@ import (
 	"github.com/gin-gonic/gin"
 	"github.com/gutakk/go-google-scraper/db"
 	"github.com/gutakk/go-google-scraper/models"
-	"github.com/gutakk/go-google-scraper/tests"
+	testConfig "github.com/gutakk/go-google-scraper/tests/config"
+	testDB "github.com/gutakk/go-google-scraper/tests/db"
+	testHttp "github.com/gutakk/go-google-scraper/tests/http"
 	"github.com/stretchr/testify/suite"
 	"gopkg.in/go-playground/assert.v1"
 	"gorm.io/driver/postgres"
@@ -28,14 +30,14 @@ type RegisterDbTestSuite struct {
 }
 
 func (s *RegisterDbTestSuite) SetupTest() {
-	testDB, _ := gorm.Open(postgres.Open(tests.ConstructTestDsn()), &gorm.Config{})
+	testDB, _ := gorm.Open(postgres.Open(testDB.ConstructTestDsn()), &gorm.Config{})
 	db.GetDB = func() *gorm.DB {
 		return testDB
 	}
 
 	_ = db.GetDB().AutoMigrate(&models.User{})
 
-	s.engine = tests.GetRouter(true)
+	s.engine = testConfig.GetRouter(true)
 	new(RegisterController).applyRoutes(EnsureGuestUserGroup(s.engine))
 
 	s.headers = http.Header{}
@@ -59,7 +61,7 @@ func TestRegisterDbTestSuite(t *testing.T) {
 }
 
 func (s *RegisterDbTestSuite) TestRegisterWithValidParameters() {
-	response := tests.PerformRequest(s.engine, "POST", "/register", s.headers, s.formData)
+	response := testHttp.PerformRequest(s.engine, "POST", "/register", s.headers, s.formData)
 
 	assert.Equal(s.T(), http.StatusFound, response.Code)
 	assert.Equal(s.T(), "/login", response.Header().Get("Location"))
@@ -68,7 +70,7 @@ func (s *RegisterDbTestSuite) TestRegisterWithValidParameters() {
 func (s *RegisterDbTestSuite) TestRegisterWithBlankEmailValidation() {
 	s.formData.Del("email")
 
-	response := tests.PerformRequest(s.engine, "POST", "/register", s.headers, s.formData)
+	response := testHttp.PerformRequest(s.engine, "POST", "/register", s.headers, s.formData)
 	p, err := ioutil.ReadAll(response.Body)
 	pageError := err == nil && strings.Index(string(p), "Invalid email format") > 0
 
@@ -79,7 +81,7 @@ func (s *RegisterDbTestSuite) TestRegisterWithBlankEmailValidation() {
 func (s *RegisterDbTestSuite) TestRegisterWithBlankPasswordValidation() {
 	s.formData.Del("password")
 
-	response := tests.PerformRequest(s.engine, "POST", "/register", s.headers, s.formData)
+	response := testHttp.PerformRequest(s.engine, "POST", "/register", s.headers, s.formData)
 	p, err := ioutil.ReadAll(response.Body)
 	pageError := err == nil && strings.Index(string(p), "Password is required") > 0
 	isEmailFieldValueExist := err == nil && strings.Index(string(p), s.email) > 0
@@ -92,7 +94,7 @@ func (s *RegisterDbTestSuite) TestRegisterWithBlankPasswordValidation() {
 func (s *RegisterDbTestSuite) TestRegisterWithPasswordNotMatchValidation() {
 	s.formData.Set("confirm-password", "invalid")
 
-	response := tests.PerformRequest(s.engine, "POST", "/register", s.headers, s.formData)
+	response := testHttp.PerformRequest(s.engine, "POST", "/register", s.headers, s.formData)
 	p, err := ioutil.ReadAll(response.Body)
 	pageError := err == nil && strings.Index(string(p), "Passwords do not match") > 0
 	isEmailFieldValueExist := err == nil && strings.Index(string(p), s.email) > 0
@@ -106,7 +108,7 @@ func (s *RegisterDbTestSuite) TestRegisterWithTooShortPasswordValidation() {
 	s.formData.Set("password", "12345")
 	s.formData.Set("confirm-password", "12345")
 
-	response := tests.PerformRequest(s.engine, "POST", "/register", s.headers, s.formData)
+	response := testHttp.PerformRequest(s.engine, "POST", "/register", s.headers, s.formData)
 	p, err := ioutil.ReadAll(response.Body)
 	pageError := err == nil && strings.Index(string(p), "Password must be longer than 6") > 0
 	isEmailFieldValueExist := err == nil && strings.Index(string(p), s.email) > 0
@@ -121,17 +123,17 @@ func (s *RegisterDbTestSuite) TestDisplayRegisterWithAuthenticatedUser() {
 	cookie := "go-google-scraper=MTYwNjQ2Mjk3MXxEdi1CQkFFQ180SUFBUkFCRUFBQUlmLUNBQUVHYzNSeWFXNW5EQWtBQjNWelpYSmZhV1FFZFdsdWRBWUVBUDRFdFE9PXzl6APqAQw3gAQqlHoXMYrPpnqPFkEP8SRHJZEpl-_LDQ=="
 	s.headers.Set("Cookie", cookie)
 
-	response := tests.PerformRequest(s.engine, "GET", "/register", s.headers, nil)
+	response := testHttp.PerformRequest(s.engine, "GET", "/register", s.headers, nil)
 
 	assert.Equal(s.T(), http.StatusFound, response.Code)
 	assert.Equal(s.T(), "/", response.Header().Get("Location"))
 }
 
 func TestDisplayRegister(t *testing.T) {
-	engine := tests.GetRouter(true)
+	engine := testConfig.GetRouter(true)
 	new(RegisterController).applyRoutes(EnsureGuestUserGroup(engine))
 
-	response := tests.PerformRequest(engine, "GET", "/register", nil, nil)
+	response := testHttp.PerformRequest(engine, "GET", "/register", nil, nil)
 	p, err := ioutil.ReadAll(response.Body)
 	pageOK := err == nil && strings.Index(string(p), "<title>Register</title>") > 0
